requirements.txt
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.2.1
openai==1.51.2
Flask==3.1.2
nltk==3.9.2
scikit-learn==1.8.0
scipy==1.17.0
numpy==2.4.1
